Add files via upload
These are the second and third versions of my game. the second version has not been fully test yet, mainly in relation to testing if the third and fourth hand situations fully work. The second file is the third version of the project, which i made because the second version was in my opinion, overengineered, with the player class doing way to much, and far to many repeating checks, so i tried to condense down a lot of the code in the third version, which seems to have mostly succeeded.  I intend to make a fourth version at some point to add more then one player to the game, even if it would just be the one person controlling them, and at some point adding a gui with cards to the game.